fixed save and resume model

+++ main.py
@@ -4,6 +4,7 @@ import numpy as np
 from selenium import webdriver
 from webdriver_manager.chrome import ChromeDriverManager
 from selenium.webdriver.common.by import By
+from selenium.webdriver.common.keys import Keys
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.common.action_chains import ActionChains
@@ -15,25 +16,41 @@ from helpers import *
 from time import sleep
 
 # Your starting YouTube video URL
-video_url = 'https://www.youtube.com/watch?v=Zr2TXJ1Gfh8'#input("Enter starting youtube video: ")
+
 
 adblockextension='/home/zorin/.var/app/com.google.Chrome/config/google-chrome/Default/Extensions/cjpalhdlnbpafiamejdnhcphjbkeiagm/1.56.0_0'
 userprofile='/home/zorin/.var/app/com.google.Chrome/config/google-chrome/Default'
 
+thumb_width=320
+thumb_height=180
 
+cleanstart=False
 
-cur=curiosity(False,320,180)
+# Define the path to your CSV file
+csv_file_path = 'saved_videos.csv'
+
+##########################################################
+
+if cleanstart:
+    video_url = input("Enter starting youtube video: ")
+else:
+    try:
+        video_url = get_last_video_from_csv(csv_file_path)
+    except:
+        video_url = input("Enter starting youtube video: ")
+
+cur=curiosity(not cleanstart,thumb_width,thumb_height)
 
 # Create the 'thumbs' directory if it doesn't exist
 if not os.path.exists('thumbs'):
     os.makedirs('thumbs')
 
-# Define the path to your CSV file
-csv_file_path = 'saved_videos.csv'
 
-# Check if the file exists and delete it if it does
-if os.path.exists(csv_file_path):
-    os.remove(csv_file_path)
+
+if cleanstart:
+    # Check if the file exists and delete it if it does
+    if os.path.exists(csv_file_path):
+        os.remove(csv_file_path)
 
 # Setting up the WebDriver with webdriver-manager
 options = webdriver.ChromeOptions()
@@ -71,18 +88,28 @@ try:
     except Exception as e:
         print("No cookies popup to accept or error clicking accept:", e)
     
-    index=0
+    if cleanstart:
+        index=0
+    else:
+        try:
+            index=count_lines_in_csv(csv_file_path)
+        except:
+            index=0
+
     while True:
         print("---start---",video_url)
         driver.get(video_url)
         sleep(1)
         # Wait for the video player to be ready
         video_player = WebDriverWait(driver, 15).until(
-            EC.presence_of_element_located((By.CSS_SELECTOR, "video"))
+            EC.presence_of_element_located((By.CSS_SELECTOR, "#player"))
         )
-        
+        #driver.execute_script("document.querySelector('player').play();")
+        play_button = driver.find_element(By.CLASS_NAME, 'ytp-play-button')
+        play_button.click()
+
         # Click on the video player to start playing the video
-        ActionChains(driver).move_to_element(video_player).click().perform()
+        #ActionChains(driver).move_to_element(video_player).click().perform()
         print("Video playback started.")
         sleep(3)
         # Wait for the page to load and for thumbnails to be available
@@ -94,7 +121,6 @@ try:
         #thumbnails = driver.find_elements(By.XPATH, '//a[@id="thumbnail"][@class="yt-simple-endpoint inline-block style-scope ytd-thumbnail"]')
         thumbnails = driver.find_elements(By.ID, "thumbnail")
 
-
         video_urls = []
         video_titles=[]
         thumbnails_img= []
@@ -104,55 +130,59 @@ try:
         for i, thumbnail in enumerate(thumbnails):
             #print(thumbnail.get_attribute("outerHTML"))
             video_url = thumbnail.get_attribute("href")
-            video_url='f'https://youtube.com/watch?v={extract_video_id(video_url)}'
 
             if video_url:
-                if video_url not in all_videos and "short" not in video_url and "channel" not in video_url:
-                    video_urls.append(video_url)
-                    all_videos.append(video_url)
-                    try:
-                        title_element =thumbnail.find_element(By.XPATH, ".//ancestor::ytd-compact-video-renderer//h3")
-                        video_title = title_element.text if title_element else "No title found"
-                        print("video_title",video_title)
-                    except:
-                        video_title="No title"
-                    video_titles.append(video_title)
-                    # Sometimes, the href might not contain a direct link to the thumbnail.
-                    # This is a simple workaround to focus on those with direct image sources.
-                    try:
+                try:
+                    video_url='https://youtube.com/watch?v='+extract_video_id(video_url)
+                except:
+                    video_url=False
+                if video_url:
+                    if video_url not in all_videos and "short" not in video_url and "channel" not in video_url:
+                        video_urls.append(video_url)
+                        all_videos.append(video_url)
+                        try:
+                            title_element =thumbnail.find_element(By.XPATH, ".//ancestor::ytd-compact-video-renderer//h3")
+                            video_title = title_element.text if title_element else "No title found"
+                            print("video_title",video_title)
+                        except:
+                            video_title="No title"
+                        video_titles.append(video_title)
+                        # Sometimes, the href might not contain a direct link to the thumbnail.
+                        # This is a simple workaround to focus on those with direct image sources.
+                        try:
 
-                        #new method
-                        img_uri=f'thumbs/thumb_{i}.jpg'
-                        if (get_small_thumbnail(video_url,img_uri)):
-                            thumbnails_img.append(img_uri)
-                            image=cur.prepare_image(img_uri)
-                            mse=cur.predict_and_calculate_mse(image)
-                            mse*=1000
-                            #print("mse",mse)
-                            mse_scores.append(mse)
-                        """    
-                        img_url = thumbnail.find_element(By.TAG_NAME, "img").get_attribute("src")
-                        if img_url:
-                            response = requests.get(img_url)
-                            img = Image.open(BytesIO(response.content))
+                            #new method
                             img_uri=f'thumbs/thumb_{i}.jpg'
-                            img.save(img_uri)
-                            thumbnails_img.append(img_uri)
-                            #print(f'Saved thumb_{i}.jpg')
+                            if (get_small_thumbnail(video_url,img_uri)):
+                                thumbnails_img.append(img_uri)
+                                image=cur.prepare_image(img_uri)
+                                mse=cur.predict_and_calculate_mse(image)
+                                mse*=1000
+                                #print("mse",mse)
+                                mse_scores.append(mse)
+                            """    
+                            img_url = thumbnail.find_element(By.TAG_NAME, "img").get_attribute("src")
+                            if img_url:
+                                response = requests.get(img_url)
+                                img = Image.open(BytesIO(response.content))
+                                img_uri=f'thumbs/thumb_{i}.jpg'
+                                img.save(img_uri)
+                                thumbnails_img.append(img_uri)
+                                #print(f'Saved thumb_{i}.jpg')
 
-                            #do the image analysis
-                            image=cur.prepare_image(img_uri)
-                            mse=cur.predict_and_calculate_mse(image)
-                            mse*=1000
-                            #print("mse",mse)
-                            mse_scores.append(mse)
-                        else:
-                            print("??????????????????????")
-                            print("could not get img url for",video_title)
-                            print("¿¿¿¿¿¿¿¿¿¿¿¿¿¿¿¿¿¿¿¿¿¿¿")
-                        """
-                    except Exception as e:
-                        print(f"Error downloading thumbnail {i}: {e}")
+                                #do the image analysis
+                                image=cur.prepare_image(img_uri)
+                                mse=cur.predict_and_calculate_mse(image)
+                                mse*=1000
+                                #print("mse",mse)
+                                mse_scores.append(mse)
+                            else:
+                                print("??????????????????????")
+                                print("could not get img url for",video_title)
+                                print("¿¿¿¿¿¿¿¿¿¿¿¿¿¿¿¿¿¿¿¿¿¿¿")
+                            """
+                        except Exception as e:
+                            print(f"Error downloading thumbnail {i}: {e}")
 
         #now update the model
         for img in thumbnails_img:
@@ -160,7 +190,7 @@ try:
             cur.update_model_with_new_image(image)
         
 
-        print("num scores",mse_scores,"::::::")
+        
         #get the max curiosity score
         maxindex=mse_scores.index(max(mse_scores))
         sorted_indices = sorted(range(len(mse_scores)), key=lambda i: mse_scores[i], reverse=True)
@@ -168,7 +198,6 @@ try:
         sorted_video_urls = [video_urls[i] for i in sorted_indices]
         sorted_video_titles=[video_titles[i] for i in sorted_indices]
         
-
         #besturl=video_urls[maxindex]
         besturl=sorted_video_urls[0]
 
@@ -197,9 +226,12 @@ try:
         print("#########################################")
         print("")
 
+        #save model
+        if not cleanstart:
+            cur.save_model()
+
         
 
 finally:
-    pass
-    #driver.quit()
+    driver.quit()
